pyproject.toml
[tool.poetry]
name = "windbird"
version = "0.1.0"
description = "Windbird project"
authors = ["Your Name <you@example.com>"]

# 配置阿里云镜像源为默认源
[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
asgiref = "^3.8"
certifi = "^2025.1"
charset-normalizer = "^3.4"
django-cors-headers = "^4.7"
djangorestframework = "^3.15"
gunicorn = "^23.0"
mysqlclient = "^2.2"
requests = "^2.32"
urllib3 = "^2.3"
cos-python-sdk-v5 = "^1.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
